Move and split existing hooks
