+++ app/importer/client/components/NewImportPage.js
@@ -25,7 +25,7 @@ import { showImporterException } from '../functions/showImporterException';
 import { useEndpoint } from '../../../../client/contexts/ServerContext';
 import { Importers } from '../index';
 import { useSafely } from '../../../../client/hooks/useSafely';
-import { useFormatMemorySize } from '../../../ui/client/views/app/components/hooks';
+import { useFormatMemorySize } from '../../../../client/hooks/useFormatMemorySize';
 
 function NewImportPage() {
 	const t = useTranslation();
