+++ app/importer/client/components/PrepareImportPage.js
@@ -12,6 +12,7 @@ import {
 	Pagination,
 	Tabs,
 } from '@rocket.chat/fuselage';
+import { useDebouncedValue } from '@rocket.chat/fuselage-hooks';
 import React, { useEffect, useState, useMemo, useCallback } from 'react';
 import s from 'underscore.string';
 
@@ -31,7 +32,6 @@ import { showImporterException } from '../functions/showImporterException';
 import { useRoute } from '../../../../client/contexts/RouterContext';
 import { useSafely } from '../../../../client/hooks/useSafely';
 import { useEndpoint } from '../../../../client/contexts/ServerContext';
-import { useDebounce } from '../../../ui/client/views/app/components/hooks';
 
 const waitFor = (fn, predicate) => new Promise((resolve, reject) => {
 	const callPromise = () => {
@@ -307,7 +307,8 @@ function PrepareImportPage() {
 	const [tab, setTab] = useState('users');
 	const handleTabClick = useMemo(() => (tab) => () => setTab(tab), []);
 
-	const statusDebounced = useDebounce(status, 100);
+	const statusDebounced = useDebouncedValue(status, 100);
+
 	return <Page>
 		<Page.Header title={t('Importing_Data')}>
 			<ButtonGroup>
